fix: split packaged-reference docs tests for backend size gate
Move packaged-goal, packaged-tts, and packaged-reference markdown-link
coverage into docs_packaged_reference_test.go so docs_test.go stays under
the 1000-line backend size limit.
